# backend/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"PortfolioFit": {
				"properties": {
					"currentSectorExposure": {
						"type": "string"
					},
					"projectedSectorExposure": {
						"type": "string"
					},
					"sector": {
						"type": "string"
					},
					"overexposureWarning": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"currentSectorExposure",
					"projectedSectorExposure",
					"sector",
					"overexposureWarning",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WatchlistItemResponse": {
				"properties": {
					"symbol": {
						"type": "string"
					},
					"currentPrice": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"changePercent": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"priceSinceAdded": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"targetEntryPrice": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"stopLoss": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"notes": {
						"type": "string",
						"nullable": true
					},
					"entryZone": {
						"type": "boolean",
						"nullable": true
					},
					"stopLossBreached": {
						"type": "boolean",
						"nullable": true
					},
					"aiSuggested": {
						"properties": {
							"computedAt": {
								"type": "string"
							},
							"basis": {
								"type": "string"
							},
							"confidence": {
								"type": "string",
								"enum": [
									"LOW",
									"MEDIUM",
									"HIGH"
								]
							},
							"stopLoss": {
								"type": "number",
								"format": "double"
							},
							"takeProfit": {
								"type": "number",
								"format": "double"
							},
							"entry": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"computedAt",
							"basis",
							"confidence",
							"stopLoss",
							"takeProfit",
							"entry"
						],
						"type": "object",
						"nullable": true
					},
					"portfolioFit": {
						"allOf": [
							{
								"$ref": "#/components/schemas/PortfolioFit"
							}
						],
						"nullable": true
					},
					"addedAt": {
						"type": "string",
						"format": "date-time"
					},
					"logo": {
						"type": "string",
						"nullable": true
					},
					"alerts": {
						"items": {
							"properties": {
								"isActive": {
									"type": "boolean"
								},
								"threshold": {
									"type": "number",
									"format": "double",
									"nullable": true
								},
								"type": {
									"type": "string"
								},
								"id": {
									"type": "string"
								}
							},
							"required": [
								"isActive",
								"threshold",
								"type",
								"id"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"symbol",
					"currentPrice",
					"changePercent",
					"priceSinceAdded",
					"targetEntryPrice",
					"stopLoss",
					"notes",
					"entryZone",
					"stopLossBreached",
					"aiSuggested",
					"portfolioFit",
					"addedAt",
					"logo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WatchlistDataResponse_WatchlistItemResponse-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/WatchlistItemResponse"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WatchlistDataResponse_WatchlistItemResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/WatchlistItemResponse"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddToWatchlistBody": {
				"properties": {
					"symbol": {
						"type": "string"
					},
					"targetEntryPrice": {
						"type": "number",
						"format": "double"
					},
					"stopLoss": {
						"type": "number",
						"format": "double"
					},
					"notes": {
						"type": "string"
					}
				},
				"required": [
					"symbol"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateWatchlistBody": {
				"properties": {
					"targetEntryPrice": {
						"type": "number",
						"format": "double"
					},
					"stopLoss": {
						"type": "number",
						"format": "double"
					},
					"notes": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"WatchlistDataResponse_void_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WatchlistDataResponse_any_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ConvertToPositionBody": {
				"properties": {
					"entryPrice": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"AlertType": {
				"$ref": "#/components/schemas/_36_Enums.AlertType"
			},
			"CreateAlertBody": {
				"properties": {
					"type": {
						"$ref": "#/components/schemas/AlertType"
					},
					"threshold": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateAlertBody": {
				"properties": {
					"threshold": {
						"type": "number",
						"format": "double"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"NotificationItem": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"body": {
						"type": "string"
					},
					"read": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"title",
					"body",
					"read",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetNotificationsEnvelopeResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/NotificationItem"
						},
						"type": "array"
					},
					"nextCursor": {
						"type": "string",
						"nullable": true
					},
					"hasMore": {
						"type": "boolean"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"nextCursor",
					"hasMore",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NotificationSummary": {
				"properties": {
					"unreadCount": {
						"type": "number",
						"format": "double"
					},
					"preview": {
						"items": {
							"$ref": "#/components/schemas/NotificationItem"
						},
						"type": "array"
					}
				},
				"required": [
					"unreadCount",
					"preview"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NotificationDataResponse_NotificationSummary_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/NotificationSummary"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NotificationsBatchUpdateResult": {
				"properties": {
					"updated": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"updated"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NotificationDataResponse_NotificationsBatchUpdateResult_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/NotificationsBatchUpdateResult"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MarkMultipleNotificationsReadBody": {
				"properties": {
					"notificationIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"notificationIds"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NotificationDataResponse_NotificationItem_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/NotificationItem"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SimpleNotificationMessageResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"MarkReadResult": {
				"properties": {
					"updated": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"properties": {
								"isRead": {
									"type": "boolean"
								},
								"id": {
									"type": "string"
								}
							},
							"required": [
								"isRead",
								"id"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"updated"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NewsDataResponse_MarkReadResult_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/MarkReadResult"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MarkMultipleReadBody": {
				"properties": {
					"articleIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"articleIds"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ArticleActionState": {
				"properties": {
					"id": {
						"type": "string"
					},
					"isRead": {
						"type": "boolean"
					},
					"isSaved": {
						"type": "boolean"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NewsDataResponse_ArticleActionState_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/ArticleActionState"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "stock-backend",
		"version": "1.0.0",
		"description": "Stock prediction, decision making and analytics platform.",
		"contact": {
			"name": "Hamza Ahmed"
		}
	},
	"paths": {
		"/api/v1/watchlist": {
			"get": {
				"operationId": "GetWatchlist",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WatchlistDataResponse_WatchlistItemResponse-Array_"
								}
							}
						}
					}
				},
				"description": "Get all watchlist items for the authenticated user with real-time pricing and AI zones.",
				"tags": [
					"Watchlist"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "AddToWatchlist",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WatchlistDataResponse_WatchlistItemResponse_"
								}
							}
						}
					}
				},
				"description": "Add a stock symbol to user's watchlist.",
				"tags": [
					"Watchlist"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddToWatchlistBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/watchlist/{symbol}": {
			"patch": {
				"operationId": "UpdateWatchlistEntry",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WatchlistDataResponse_WatchlistItemResponse_"
								}
							}
						}
					}
				},
				"description": "Update target prices or notes for a watchlist entry.",
				"tags": [
					"Watchlist"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "symbol",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateWatchlistBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "RemoveFromWatchlist",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WatchlistDataResponse_void_"
								}
							}
						}
					}
				},
				"description": "Remove a stock from user's watchlist.",
				"tags": [
					"Watchlist"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "symbol",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/watchlist/{symbol}/convert-to-position": {
			"post": {
				"operationId": "ConvertToPosition",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WatchlistDataResponse_any_"
								}
							}
						}
					}
				},
				"description": "Convert a watchlist entry to a portfolio position.",
				"tags": [
					"Watchlist"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "symbol",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConvertToPositionBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/watchlist/{symbol}/alerts": {
			"get": {
				"operationId": "GetAlerts",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WatchlistDataResponse_any_"
								}
							}
						}
					}
				},
				"description": "Get all active alerts for a watchlist symbol.",
				"tags": [
					"Watchlist"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "symbol",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateAlert",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WatchlistDataResponse_any_"
								}
							}
						}
					}
				},
				"description": "Create an alert rule for a watchlist symbol.",
				"tags": [
					"Watchlist"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "symbol",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAlertBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/watchlist/{symbol}/alerts/{id}": {
			"patch": {
				"operationId": "UpdateAlert",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WatchlistDataResponse_any_"
								}
							}
						}
					}
				},
				"description": "Update an alert rule.",
				"tags": [
					"Watchlist"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "symbol",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAlertBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAlert",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WatchlistDataResponse_void_"
								}
							}
						}
					}
				},
				"description": "Delete an alert rule.",
				"tags": [
					"Watchlist"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "symbol",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/v1/notifications": {
			"get": {
				"operationId": "GetNotifications",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetNotificationsEnvelopeResponse"
								}
							}
						}
					}
				},
				"description": "Get paginated notifications for the authenticated user.",
				"tags": [
					"Notifications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/notifications/summary": {
			"get": {
				"operationId": "GetNotificationSummary",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotificationDataResponse_NotificationSummary_"
								}
							}
						}
					}
				},
				"description": "Get notification summary (unread count + latest 5 notifications).",
				"tags": [
					"Notifications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/notifications/read-all": {
			"patch": {
				"operationId": "MarkAllAsRead",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotificationDataResponse_NotificationsBatchUpdateResult_"
								}
							}
						}
					}
				},
				"description": "Mark all unread notifications as read.",
				"tags": [
					"Notifications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/notifications/read-multiple": {
			"patch": {
				"operationId": "MarkMultipleAsRead",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotificationDataResponse_NotificationsBatchUpdateResult_"
								}
							}
						}
					}
				},
				"description": "Mark multiple notifications as read.",
				"tags": [
					"Notifications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MarkMultipleNotificationsReadBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/notifications/{id}/read": {
			"patch": {
				"operationId": "MarkAsRead",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotificationDataResponse_NotificationItem_"
								}
							}
						}
					}
				},
				"description": "Mark a single notification as read.",
				"tags": [
					"Notifications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/notifications/{id}": {
			"delete": {
				"operationId": "DeleteNotification",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SimpleNotificationMessageResponse"
								}
							}
						}
					}
				},
				"description": "Delete a notification.",
				"tags": [
					"Notifications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/v1/news/read-all": {
			"patch": {
				"operationId": "MarkAllAsRead",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NewsDataResponse_MarkReadResult_"
								}
							}
						}
					}
				},
				"description": "Mark all news articles as read for the user.",
				"tags": [
					"News"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/news/read-multiple": {
			"patch": {
				"operationId": "MarkMultipleAsRead",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NewsDataResponse_MarkReadResult_"
								}
							}
						}
					}
				},
				"description": "Mark multiple news articles as read.",
				"tags": [
					"News"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MarkMultipleReadBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/news/{id}/read": {
			"patch": {
				"operationId": "MarkAsRead",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NewsDataResponse_ArticleActionState_"
								}
							}
						}
					}
				},
				"description": "Mark a single news article as read.",
				"tags": [
					"News"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}